earthpulse-backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.11
alembic==1.14.0

# Scraping
beautifulsoup4==4.12.3
requests==2.32.3
selenium==4.26.1
scrapy==2.12.0
newspaper3k==0.2.8

# NLP
spacy==3.8.0
transformers==4.48.1
torch==2.5.1+cpu
geopy==2.4.1

# Geospatial
geopandas==1.0.1
shapely==2.0.6
folium==0.19.1

# Satellite & Image Processing
sentinelhub==3.11.0
earthengine-api==1.2.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
rasterio==1.4.3
opencv-python==4.10.0.84
Pillow==11.1.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
celery==5.4.0
redis==5.0.8
httpx==0.28.1